requirements.txt
requests>=2.25.0
Pillow>=10.4.0